split parser keyword spelling
